Changed to Int from Short

+++ app/src/main/java/net/gas/gascontact/business/database/entities/Posts.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "posts")
 data class Posts (
-    @PrimaryKey val id: Short,
+    @PrimaryKey val id: Int,
     val name: String?,
     val rangir: Int?
 )
